refactor(operator): extract Kafka wait helpers into ITBase
Move duplicated Kafka broker readiness and registry deployment
wait logic from 4 test classes into shared ITBase helpers with
named timeout constants (KAFKA_BROKER_READY_TIMEOUT = 10min,
KAFKA_REGISTRY_READY_TIMEOUT = 8min). Removes ~84 lines of
duplicated await() boilerplate.
