+++ operator/controller/src/test/java/io/apicurio/registry/operator/it/KafkaSqlOAuthITTest.java
@@ -1,7 +1,6 @@
 package io.apicurio.registry.operator.it;
 
 import io.apicurio.registry.operator.api.v1.ApicurioRegistry3;
-import io.fabric8.kubernetes.api.model.PodCondition;
 import io.quarkus.test.junit.QuarkusTest;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Tag;
@@ -13,9 +12,6 @@ import static io.apicurio.registry.operator.Tags.AUTH;
 import static io.apicurio.registry.operator.Tags.KAFKA;
 import static io.apicurio.registry.operator.Tags.SLOW;
 import static io.apicurio.registry.operator.resource.ResourceFactory.deserialize;
-import static java.time.Duration.ofMinutes;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.awaitility.Awaitility.await;
 
 @QuarkusTest
 @Tag(KAFKA)
@@ -40,11 +36,7 @@ public class KafkaSqlOAuthITTest extends BaseAuthITTest {
                 .create();
         final var clusterName = "oauth-example-cluster";
 
-        await().atMost(ofMinutes(10)).ignoreExceptions().untilAsserted(() ->
-                // Strimzi uses StrimziPodSet instead of ReplicaSet, so we have to check pods
-                assertThat(client.pods().inNamespace(namespace).withName(clusterName + "-dual-role-0").get().getStatus()
-                        .getConditions()).filteredOn(c -> "Ready".equals(c.getType())).map(PodCondition::getStatus)
-                        .containsOnly("True"));
+        waitForKafkaBrokerReady(clusterName);
 
         // We're guessing the value here to avoid using Strimzi Java model, and relying on retries below.
         var bootstrapServers = clusterName + "-kafka-bootstrap." + namespace + ".svc:9093";
@@ -57,17 +49,6 @@ public class KafkaSqlOAuthITTest extends BaseAuthITTest {
 
         client.resource(registry).create();
 
-        await().atMost(ofMinutes(8)).ignoreExceptions().until(() -> {
-            assertThat(client.apps().deployments().inNamespace(namespace)
-                    .withName(registry.getMetadata().getName() + "-app-deployment").get().getStatus()
-                    .getReadyReplicas().intValue()).isEqualTo(1);
-            var podName = client.pods().inNamespace(namespace).list().getItems().stream()
-                    .map(pod -> pod.getMetadata().getName())
-                    .filter(podN -> podN.startsWith(registry.getMetadata().getName() + "-app-deployment"))
-                    .findFirst().get();
-            assertThat(client.pods().inNamespace(namespace).withName(podName).getLog())
-                    .contains("Using Kafka-SQL artifactStore");
-            return true;
-        });
+        waitForKafkaSqlRegistryReady(registry);
     }
 }
